updated wordleStats.h
struct of win, lose, games played, win percent

+++ wordleStats.h
@@ -6,7 +6,7 @@ typedef struct{
 	int win;
 	int lose;
 	int totalGames;
-	int winPercent;
+	double winPercent;
 
 } Gamestats;
